refactor: 재고 이중쓰기 구조 제거 및 Kafka 비동기 처리 도입
feat: 재고 관리 개선 - Redis Lua 스크립트 및 Kafka 연동

chore: Redis Sentinel 구성 반영 및 재고 로직 분리

+++ src/main/java/org/allmart/productservice/application/port/iin/ProductUseCase.java
@@ -9,7 +9,7 @@ public interface ProductUseCase {
 
     Product registerProduct(Product product);
 
-    Product decreaseStock(String productId, int quantity);
+    Product decreaseStock(String productId, long quantity);
 
-    Product increaseStock(String productId, int quantity);
+    Product increaseStock(String productId, long quantity);
 }
